Progetto/Notifier/requirements.txt
Flask==3.0.0
psutil==5.9.5
pyOpenSSL==23.3.0
Requests==2.31.0
tenacity==8.2.3